Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    int a;
-    int b;
-    int c;
-    int average;
+    float a;
+    float b;
+    float c;
+    float average;
     scanf("%f",&a);
     scanf("%f",&b);
     scanf("%f",&c);
